351 renamed EnclaveResolver to TenantResolver
