+++ extensions/servlet/src/main/resources/com/stormpath/sdk/servlet/config/web.stormpath.properties
@@ -185,7 +185,7 @@ stormpath.web.idSite.OrganizationResolverFactory = com.stormpath.sdk.servlet.fil
 # Inferred based on heuristics by default.  However if your application is not deployed to an apex domain, like
 # myapp.com, you *must* specify your application's base domain, e.g. myapp.mycompany.com
 stormpath.web.application.domain =
-stormpath.web.application.enclave.resolver = com.stormpath.sdk.servlet.enclave.DefaultOrganizationEnclaveResolverFactory
+stormpath.web.application.tenant.resolver = com.stormpath.sdk.servlet.tenant.DefaultOrganizationTenantResolverFactory
 
 stormpath.web.request.event.publisher = com.stormpath.sdk.servlet.event.impl.EventPublisherFactory
 stormpath.web.request.event.listener = com.stormpath.sdk.servlet.event.RequestEventListenerAdapter
